requirements.txt
azure-functions
